backend/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.4.1
yfinance==0.2.37
pandas==2.2.1
python-multipart==0.0.9
sqlalchemy==2.0.29
httpx==0.27.0
kiteconnect
pyotp
playwright
setuptools
openpyxl
